Final update! Done: - So much that I even don't remember XD
Game isn't what it suppose to be but when time is so short you need to do something.
And gamedev always finds the way XD
There was a lot of things that i wanted to add but in like month time span I have done quite a lot (not even month and only days before uploads becouse of lots of student projects to finish)
True value that i've got from this project is that lack of time to develop might actualy make you game way diffrent then others (sometimes better sometimes worse).
Lastly NEVER RUSH PROJECTS!!! (i'm very upsed and depresed that I have to show game in this state, but after all it's not that bad I think XD)